ai-service-python/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ── Embeddings & Vector DB ─────────────────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4

# ── Data processing ────────────────────────────────────────────────────────────
pandas==2.2.2

# ── LLM ───────────────────────────────────────────────────────────────────────
openai==1.35.3

# ── LangChain / LangGraph (compatible versions) ────────────────────────────────
langchain==0.2.6
langchain-core==0.2.22
langchain-openai==0.1.14
langgraph==0.1.14

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
